Resize find size function and create new file find_size_env_var.c
