+++ Makefile
@@ -25,7 +25,8 @@ SRCS	:=	./src/main.c \
 			./src/trans_env.c \
 			./src/split_input.c \
 			./src/check_first_argv.c \
-			./src/free_functions.c
+			./src/free_functions.c \
+			./src/find_size_env_var.c
 			
 OBJS	:= ${SRCS:.c=.o}
